// src/types/share/messageData.d.ts
export interface MessageData {
    nombre_cliente: string | null
    dni: string | null
    localidad: string | null
    direccion: string | null
    asunto: string | null
    email: string | null
    mensaje_cliente: string | null
    estado: string
    celular: string | null
    archivos_adjunto: number[]
  }